Splash Screen ready
I would not merge this with other branches since it contains changes in the code in terms of there were 4 options in the main menu, now I made it that there are only two. Other than that the main letters are changed, they are animated with endlessly looping up and down animation. The highscore things are also hidden from the splash screen for now.